backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "motor>=3.7.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "tinydb>=4.8.2",
    "aiosmtplib>=3.0.2",
    "email-validator>=2.1.1",
    "mailtrap>=2.5.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
